+++ .vscode/settings.json
@@ -1,6 +1,4 @@
 {
-  // not to format on save, but instead fix ESLint errors on save
-  "editor.formatOnSave": false,
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
   }
