+++ atoms/layout/index.ts
@@ -1,3 +1,4 @@
 export { default as Layout } from './Layout';
 export { default as Header } from './header';
 export { default as Container } from './container';
+export { default as Section } from './Section';
