--- atoms/typography/Typograpy.module.scss
@@ -1,6 +1,0 @@
-@use 'color';
-@use 'font';
-.title {
-    @include font.set(20px, bold);
-    color: color.$primary;
-}
