--- atoms/typography/index.ts
@@ -1,1 +1,0 @@
-export { default as Typograpy } from './Typograpy';
