--- molecules/featured/index.ts
@@ -1,1 +1,0 @@
-export { default as Featured } from './Featured';
