+++ molecules/preview/FeaturedSection.tsx
@@ -1,0 +1,9 @@
+import { Section } from '~/atoms/layout';
+
+export default function FeaturedSection() {
+  return (
+    <Section>
+      <Section.Title>Featured</Section.Title>
+    </Section>
+  );
+}
